backend/requirements.txt
fastapi[standard]==0.129.0
beanie==1.30.0
motor==3.7.1
pydantic-settings==2.12.0
openai==1.108.1
PyJWT==2.10.1
loguru==0.7.2
slowapi==0.1.9
